EEB7: Reshuffle files and rename to be more suggestive of the direction we are taking
