+++ frontend/src/components/filtering/by-simple-search-query.vue
