+++ frontend/src/components/review-service-info/review-service-summary-graph.vue
